+++ src/domain/entities/relation.py
@@ -88,3 +88,5 @@ class Relation:
         return (self.source_entity_id, self.relation_type, self.target_entity_id)
 
 
+
+