+++ src/domain/value_objects/predefined_patterns.py
@@ -79,3 +79,5 @@ FILE_PAGE_CHUNK_PATTERN = GraphPattern(
 
 
 
+
+