+++ src/infrastructure/services/gds_service.py
@@ -181,3 +181,5 @@ class GDSService:
             self._driver = None
 
 
+
+